requirements.txt
pandas==1.5.1
scipy==1.9.3
plotly==5.11.0
nbformat==5.7.0
ipython==8.6.0
pytest-cov==4.0.0
black==22.10.0
pylint==2.15.5
isort==5.10.1
pre-commit==2.20.0